# DebugBuild and faster version