```
goos: darwin
goarch: arm64
pkg: github.com/cristalhq/builq
                 │   old.txt    │               new.txt                │
                 │    sec/op    │    sec/op     vs base                │
BuildSimple-10      499.0n ± 0%    297.6n ± 2%  -40.34% (p=0.000 n=10)
BuildManyArgs-10   1964.5n ± 1%    661.6n ± 0%  -66.32% (p=0.000 n=10)
BuildCached-10     0.6219n ± 0%   0.6266n ± 1%   +0.76% (p=0.004 n=10)
geomean             84.79n         49.79n       -41.28%

                 │    old.txt     │                new.txt                 │
                 │      B/op      │     B/op      vs base                  │
BuildSimple-10       492.0 ± 0%       864.0 ± 0%  +75.61% (p=0.000 n=10)
BuildManyArgs-10   1.040Ki ± 0%     1.188Ki ± 0%  +14.18% (p=0.000 n=10)
BuildCached-10       0.000 ± 0%       0.000 ± 0%        ~ (p=1.000 n=10) ¹
geomean                         ²                 +26.10%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                 │   old.txt    │               new.txt                │
                 │  allocs/op   │ allocs/op   vs base                  │
BuildSimple-10     8.000 ± 0%     7.000 ± 0%  -12.50% (p=0.000 n=10)
BuildManyArgs-10   24.00 ± 0%     14.00 ± 0%  -41.67% (p=0.000 n=10)
BuildCached-10     0.000 ± 0%     0.000 ± 0%        ~ (p=1.000 n=10) ¹
geomean                       ²               -20.08%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean
```